+++ scripts/01_train_test_split.py
@@ -82,7 +82,7 @@ if __name__ == "__main__":
 
     # Splits
     try:
-        cv = StratifiedGroupKFold(n_splits=args.n_splits, shuffle=True, random_state=args.seed)
+        cv = StratifiedGroupKFold(n_splits=args.n_splits, shuffle=True, random_state=args.random_state)
 
         # Try to calculate the splits for exception catchment
         next(iter(cv.split(df, df[args.target_col], df[args.group_col])))